feat: separate toolbar component
